Kinda Cool Stuff
So, I found that (using only division and multiplication of pairs of attributes), by far the most important transformation is adding an attribute With numerator 3 and denominator 4
The error in the training data is : 0.7889. This is total rooms/total bathrooms, which seems odd, but who knows. Oh actually thinking about it it seems to make sense. I will add this attribute to main.